insights: extract a store
Release note: None
